+++ HAPI_Start/src/Utilities/Math.cpp
@@ -1,6 +1,6 @@
 #include "Math.h"
 
-Vector2f Math::getDirection(Vector2i currentPosition, Vector2i destination)
+Vector2f Math::getDirection(Vector2f currentPosition, Vector2f destination)
 {
 	Vector2f vect = Vector2f(destination.x, destination.y) - Vector2f(currentPosition.x, currentPosition.y);
 	float length = sqrt(vect.x * vect.x + vect.y * vect.y);
@@ -11,9 +11,9 @@ Vector2f Math::getDirection(Vector2i currentPosition, Vector2i destination)
 	return vect;
 }
 
-bool Math::isWithinRange(Vector2i currentPosition, Vector2i targetPosition, float attackRange)
+bool Math::isWithinRange(Vector2f currentPosition, Vector2f targetPosition, float attackRange)
 {
-	Vector2i vect = targetPosition - currentPosition;
+	Vector2f vect = targetPosition - currentPosition;
 	float length = sqrt(vect.x * vect.x + vect.y * vect.y);
 	
 	return length <= attackRange;
